Fix some size_t cast warnings (on 64bit)

+++ src/Win/DlgOptions.h
@@ -48,10 +48,10 @@ class CDlgOptionsProgram;
 class CDlgOptions : public wxDialog
 {
 public:
-	static size_t GetProgramPage()	{return 0;}
-	static size_t GetFilterPage()	{return 1;}
-	static size_t GetCalendarPage()	{return 2;}
-	static size_t GetPrintPage()	{return 3;}
+	static int GetProgramPage()		{return 0;}
+	static int GetFilterPage()		{return 1;}
+	static int GetCalendarPage()	{return 2;}
+	static int GetPrintPage()		{return 3;}
 
 	CDlgOptions(
 			CAgilityBookDoc* pDoc,
